Menu Creation
Improvement adjustments over Creation of OtherWorker. Started Alter Other Worker.

+++ NTeam/Classes/SoccerPlayer.cpp
@@ -5,7 +5,7 @@
 #include "SoccerPlayer.h"
 
 
-SoccerPlayer::SoccerPlayer(unsigned int id, std::string name, std::string role, Date birthDate, int salary, std::string position, std::string club,
+SoccerPlayer::SoccerPlayer(unsigned int id, std::string name, std::string role, Date birthDate, float salary, std::string position, std::string club,
                            unsigned short weight, unsigned short height, unsigned long marketPrice,
                            unsigned int daysActive, bool injured): Staff(id, role, name, birthDate),
                            position(position), club(club), weight(weight), height(height), marketPrice(marketPrice),
